+++ paq8pxd.cpp
@@ -127,6 +127,7 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/mft.hpp"
 #include "prt/contextmap2.hpp"
 #include "prt/contextmap.hpp"
+#include "prt/bh.hpp"
 
 #include "models/model.hpp"
 #include "models/sparsematch.hpp"
@@ -137,6 +138,9 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "models/dmc.hpp"
 #include "models/sparsex.hpp"
 #include "models/jpeg.hpp"
+#include "models/match1.hpp"
+#include "models/match2.hpp"
+#include "models/normal.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -258,123 +262,6 @@ int n0n1[256]; // for contectmap
 #define PNGFlag (1<<31)
 #define GrayFlag (1<<30)
 
-///////////////////////////// BH ////////////////////////////////
-
-// A BH maps a 32 bit hash to an array of B bytes (checksum and B-2 values)
-//
-// BH bh(N); creates N element table with B bytes each.
-//   N must be a power of 2.  The first byte of each element is
-//   reserved for a checksum to detect collisions.  The remaining
-//   B-1 bytes are values, prioritized by the first value.  This
-//   byte is 0 to mark an unused element.
-//
-// bh[i] returns a pointer to the i'th element, such that
-//   bh[i][0] is a checksum of i, bh[i][1] is the priority, and
-//   bh[i][2..B-1] are other values (0-255).
-//   The low lg(n) bits as an index into the table.
-//   If a collision is detected, up to M nearby locations in the same
-//   cache line are tested and the first matching checksum or
-//   empty element is returned.
-//   If no match or empty element is found, then the lowest priority
-//   element is replaced.
-
-// 2 byte checksum with LRU replacement (except last 2 by priority)
-template <U32 B> class BH {
-  enum {M=4};  // search limit
-  Array<U8, 64> t; // elements
-  //Array<U8> tmp;
-  U8 tmp[B];
-  U32 n; // size-1
-public:
-  BH(U32 i): t(i*B), n(i-1) {
-    //printf("BH %0.0f, i %d B %d power %d\n",(i*B)+0.0,i,B,(i&(i-1))==0);
-    assert(B>=2 && i>0 && (i&(i-1))==0); // size a power of 2?
-    
-  }
-  U8* operator[](U32 i);
-};
-
-template <U32 B>
-inline  U8* BH<B>::operator[](U32 i) {
-  U16 chk=(i>>16^i)&0xffff;
-  i=i*M&n;
-  U8 *p;
-  U16 *cp;
-  int j;
-  for (j=0; j<M; ++j) {
-    p=&t[(i+j)*B];
-    cp=(U16*)p;
-    if (p[2]==0) {*cp=chk;break;}
-    if (*cp==chk) break;  // found
-  }
-  if (j==0) return p+1;  // front
-  //static U8 tmp[B];  // element to move to front
-  if (j==M) {
-    --j;
-    memset(&tmp, 0, B);
-    memmove(&tmp, &chk, 2);
-    if (M>2 && t[(i+j)*B+2]>t[(i+j-1)*B+2]) --j;
-  }
-  else memcpy(&tmp, cp, B);
-  memmove(&t[(i+1)*B], &t[i*B], j*B);
-  memcpy(&t[i*B], &tmp, B);
-  return &t[i*B+1];
-}
- 
-/////////////////////////// ContextMap /////////////////////////
-//
-// A ContextMap maps contexts to a bit histories and makes predictions
-// to a Mixer.  Methods common to all classes:
-//
-// ContextMap cm(M, C); creates using about M bytes of memory (a power
-//   of 2) for C contexts.
-// cm.set(cx);  sets the next context to cx, called up to C times
-//   cx is an arbitrary 32 bit value that identifies the context.
-//   It should be called before predicting the first bit of each byte.
-// cm.mix(m) updates Mixer m with the next prediction.  Returns 1
-//   if context cx is found, else 0.  Then it extends all the contexts with
-//   global bit y.  It should be called for every bit:
-//
-//     if (bpos==0)
-//       for (int i=0; i<C; ++i) cm.set(cxt[i]);
-//     cm.mix(m);
-//
-// The different types are as follows:
-//
-// - RunContextMap.  The bit history is a count of 0-255 consecutive
-//     zeros or ones.  Uses 4 bytes per whole byte context.  C=1.
-//     The context should be a hash.
-// - SmallStationaryContextMap.  0 <= cx < M/512.
-//     The state is a 16-bit probability that is adjusted after each
-//     prediction.  C=1.
-// - ContextMap.  For large contexts, C >= 1.  Context need not be hashed.
-
-
-
-// A RunContextMap maps a context into the next byte and a repeat
-// count up to M.  Size should be a power of 2.  Memory usage is 3M/4.
-class RunContextMap {
-  BH<4> t;
-  U8* cp;
-  BlockData& x;
-public:
-  RunContextMap(int m,BlockData& bd): t(m/4),x(bd) {cp=t[0]+1;}
-  void set(U32 cx) {  // update count
-    if (cp[0]==0 || cp[1]!=x.buf(1)) cp[0]=1, cp[1]=x.buf(1);
-    else if (cp[0]<255) ++cp[0];
-    cp=t[cx]+1;
-  }
-  int p() {  // predict next bit
-    if ((cp[1]+256)>>(8-x.bpos)==x.c0)
-      return ((cp[1]>>(7-x.bpos)&1)*2-1)*ilog(cp[0]+1)*8;
-    else
-      return 0;
-  }
-  int mix(Mixer& m) {  // return run length
-    m.add(p());
-    return cp[0]!=0;
-  }
-};
 
 
 
@@ -2469,557 +2356,7 @@ void TextModel::SetContexts(Buf& buffer,Mixer& mixer) {
   Map.set(hash(i++, cWord->Hash[0], (Info.lastUpper<column)|((Info.lastUpper<Info.wordLength[0])<<1), min(5, Info.wordLength[0])));
 }
 
-//////////////////////////// matchModel ///////////////////////////
-class matchModel1: public Model {
-private:
-    BlockData& x;
-    Buf& buffer;
-    const U64 Size;
-  enum Parameters : U32{
-      MaxLen = 0xFFFF, // longest allowed match
-    MaxExtend = 0,   // longest match expansion allowed // warning: larger value -> slowdown
-    MinLen = 5,      // minimum required match length
-    StepSize = 2,    // additional minimum length increase per higher order hash
-    DeltaLen = 5,    // minimum length to switch to delta mode
-    NumCtxs = 3,     // number of contexts used
-    NumHashes = 3    // number of hashes used
 
-  };
-  Array<U32> Table;
-  StateMap StateMaps[NumCtxs];
-  SmallStationaryContextMap SCM[3];
-  StationaryMap Maps[3];
-  IndirectContext<U8> iCtx;
-  U32 hashes[NumHashes];
-  U32 ctx[NumCtxs];
-  U32 length;    // rebased length of match (length=1 represents the smallest accepted match length), or 0 if no match
-  U32 index;     // points to next byte of match in buffer, 0 when there is no match
-  U32 mask;
-  U8 expectedByte; // prediction is based on this byte (buffer[index]), valid only when length>0
-  bool delta;
-  void Update() {
-        delta = false;
-    if (length==0 && x.Match.bypass)
-      x.Match.bypass = false; // can quit bypass mode on byte boundary only
-    // update hashes
-    for (U32 i=0, minLen=MinLen+(NumHashes-1)*StepSize; i<NumHashes; i++, minLen-=StepSize) {
-      hashes[i] = hash(minLen);
-      for (U32 j=1; j<=minLen; j++)
-        hashes[i] = combine(hashes[i], buffer(j)<<i);
-      hashes[i]&=mask;
-    }
-    // extend current match, if available
-    if (length) {
-      index++;
-      if (length<MaxLen)
-        length++;
-    }
-    // or find a new match, starting with the highest order hash and falling back to lower ones
-    else {
-      U32 minLen = MinLen+(NumHashes-1)*StepSize, bestLen = 0, bestIndex = 0;
-      for (U32 i=0; i<NumHashes && length<minLen; i++, minLen-=StepSize) {
-        index = Table[hashes[i]];
-        if (index>0) {
-          length = 0;
-          while (length<(minLen+MaxExtend) && buffer(length+1)==buffer[index-length-1])
-            length++;
-          if (length>bestLen) {
-            bestLen = length;
-            bestIndex = index;
-          }
-        }
-      }
-      if (bestLen>=minLen) {
-        length = bestLen-(MinLen-1); // rebase, a length of 1 actually means a length of MinLen
-        index = bestIndex;
-      }
-      else
-        length = index = 0;
-    }
-    // update position information in hashtable
-    for (U32 i=0; i<NumHashes; i++)
-      Table[hashes[i]] = x.buf.pos;
-      expectedByte = buffer[index];
-    iCtx+=x.y, iCtx=(buffer(1)<<8)|expectedByte;
-    const uint32_t lengthIlog2 = ilog2(length + 1);
-    //no match:      lengthIlog2=0
-    //length=1..2:   lengthIlog2=1
-    //length=3..6:   lengthIlog2=2
-    //length=7..14:  lengthIlog2=3
-    //length=15..30: lengthIlog2=4
-    const uint8_t length3 = min(lengthIlog2, 3); // 2 bits
-  
-    SCM[0].set(expectedByte);
-    SCM[1].set(expectedByte);
-    SCM[2].set(x.buf.pos);
-    Maps[0].set((expectedByte<<8)|buffer(1));
-    Maps[1].set(hash(expectedByte, x.c0, buffer(1), length3 ));
-    Maps[2].set(iCtx());
-    x.Match.byte = (length)?expectedByte:0;
-  }
-public:
-  bool canBypass;
-  virtual ~matchModel1(){ }
-  int inputs() {return  2+NumCtxs+3*2+3*2+2;}
-  int nets() {return 8;}
-  int netcount() {return 1;}
-  matchModel1(BlockData& bd, U32 val1=0) :
-    x(bd),buffer(bd.buf),Size(level>9?0x80000000: CMlimit(MEM()*4)),
-    Table(Size/sizeof(U32)),
-    StateMaps{56*256, 8*256*256+1, 256*256 },
-    SCM{ {8,8},{11,1},{8,8} },
-    Maps{ {16}, {22,1}, {4,1} },
-    iCtx{19,1},
-    hashes{ 0 },
-    ctx{ 0 },
-    length(0),
-    mask(Size/sizeof(U32)-1),
-    expectedByte(0),
-    delta(false),
-    canBypass(true)
-  {
-      x.Match.bypass=false;
-      x.Match.bypassprediction=2048;
-    assert((Size&(Size-1))==0);
-  }
-  int p(Mixer& m,int val1=0,int val2=0) {
-    if (x.bpos==0)
-      Update();
-   else {
-      U8 B = x.c0<<(8-x.bpos);
-      SCM[1].set((x.bpos<<8)|(expectedByte^B));
-      Maps[1].set(hash(expectedByte, x.c0, buffer(1), buffer(2), min(3,(int)ilog2(length+1))));
-      iCtx+=x.y, iCtx=(x.bpos<<16)|(buffer(1)<<8)|(expectedByte^B);
-      Maps[2].set(iCtx());
-    }
-    const int expectedBit =length != 0 ? (expectedByte>>(7-x.bpos))&1: 0;;
-
-    if(length>0) {
-      const bool isMatch = x.bpos==0 ? buffer(1)==buffer[index-1] : ((expectedByte+256)>>(8-x.bpos))==x.c0; // next bit matches the prediction?
-      if(!isMatch) {
-        delta = (length+MinLen)>DeltaLen;
-        length = 0;
-      }
-    }
-
-    if (!(canBypass && x.Match.bypass)) {
-      for (U32 i=0; i<NumCtxs; i++)
-        ctx[i] = 0;
-      if (length>0) {
-          if (length<=16)
-            ctx[0] = (length-1)*2 + expectedBit; // 0..31
-          else
-            ctx[0] = 24 + (min(length-1, 63)>>2)*2 + expectedBit; // 32..55
-          ctx[0] = ((ctx[0]<<8) | x.c0);
-          ctx[1] = ((expectedByte<<11) | (x.bpos<<8) | buffer(1)) + 1;
-          const int sign = 2*expectedBit-1;
-          m.add(sign*(min(length,32)<<5)); // +/- 32..1024
-          m.add(sign*(ilog(length)<<2));   // +/-  0..1024
-        }
-        else { // no match at all or delta mode
-          m.add(0);
-          m.add(0);
-        }
-
-      if (delta)
-        ctx[2] = (expectedByte<<8) | x.c0;
-
-      for (U32 i=0; i<NumCtxs; i++) {
-        const U32 c = ctx[i];
-        const U32 p = StateMaps[i].p(c,x.y);
-        if (c!=0)
-          m.add((stretch(p)+1)>>1);
-        else
-          m.add(0);
-      }
-
-      SCM[0].mix(m);
-      SCM[1].mix(m, 6);
-      SCM[2].mix(m, 5);
-      Maps[0].mix(m);
-      Maps[1].mix(m);
-      Maps[2].mix(m);
-     x.Match.length3=(min(ilog2(x.Match.length),3));
-      m.set(min( x.Match.length3 + 1, 7), 8);
-    }
-    x.Match.bypassprediction = length==0 ? 2048 : (expectedBit==0 ? 1 : 4095);
-    x.Match.length = length;
-    return ilog(length);
-  }
-  };
-  struct HashElementForMatchPositions { // sizeof(HashElementForMatchPositions) = 3*4 = 12
-  static constexpr size_t N = 3;
-  uint32_t matchPositions[N];
-  void Add(uint32_t pos) {
-    if (N > 1) {
-      memmove(&matchPositions[1], &matchPositions[0], (N - 1) * sizeof(matchPositions[0]));
-    }
-    matchPositions[0] = pos;
-  }
-};
-
-
-
-class matchModel2: public Model {
-private:
-    BlockData& x;
-    static constexpr int nCM = 2;
-    static constexpr int nST = 3;
-    static constexpr int nLSM = 1;
-    static constexpr int nSM = 1;
-    Array<HashElementForMatchPositions> hashtable;
-    StateMap stateMaps[nST];
-    ContextMap2 cm;
-    LargeStationaryMap mapL[nLSM];
-    StationaryMap map[nSM];
-    static constexpr uint32_t iCtxBits = 7;
-    IndirectContext1<uint8_t> iCtx;
-    uint32_t ctx[nST] {0};
-    const int hashBits;
-
-    static constexpr int MINLEN_RM = 5; //minimum length in recovery mode before we "fully recover"
-    static constexpr int LEN5 = 5;
-    static constexpr int LEN7 = 7;
-    static constexpr int LEN9 = 9;
-
-    struct MatchInfo {
-
-      uint32_t length = 0; /**< rebased length of match (length=1 represents the smallest accepted match length), or 0 if no match */
-      uint32_t index = 0; /**< points to next byte of match in buf, 0 when there is no match */
-      uint32_t lengthBak = 0; /**< allows match recovery after a 1-byte mismatch */
-      uint32_t indexBak = 0;
-      uint8_t expectedByte = 0; /**< prediction is based on this byte (buf[index]), valid only when length>0 */
-      bool delta = false; /**< indicates that a match has just failed (delta mode) */
-
-      bool isInNoMatchMode() {
-        return length == 0 && !delta && lengthBak == 0;
-      }
-
-      bool isInPreRecoveryMode() {
-        return length == 0 && !delta && lengthBak != 0;
-      }
-
-      bool isInRecoveryMode() {
-        return length != 0 && lengthBak != 0;
-      }
-
-      uint32_t recoveryModePos() {
-        assert(isInRecoveryMode()); //must be in recovery mode
-        assert(length - lengthBak <= MINLEN_RM);
-        return length - lengthBak;
-      }
-
-      uint64_t prio() {
-        return
-          static_cast<uint64_t>(length != 0) << 49 | //normal mode (match)
-          static_cast<uint64_t>(delta) << 48 | //delta mode
-          static_cast<uint64_t>(delta ? lengthBak : length) << 32 | //the lnger wins
-          static_cast<uint64_t>(index); //the more recent wins
-      }
-      bool isBetterThan(MatchInfo* other) {
-        return this->prio() > other->prio();
-      }
-
-      void update(BlockData& x) {
-        if (false) {
-          printf("- pos %d %d  index %d  length %d  lengthBak %d  delta %d\n", x.blpos, x.bpos, index, length, lengthBak, delta ? 1 : 0);
-        }
-        if (length != 0) {
-          const int expectedBit = (expectedByte >> ((8 - x.bpos) & 7)) & 1;
-          if (x.y != expectedBit) {
-            if (isInRecoveryMode()) { // another mismatch in recovery mode -> give up
-              lengthBak = 0;
-              indexBak = 0;
-            }
-            else { //backup match information: maybe we can recover it just after this mismatch
-              lengthBak = length;
-              indexBak = index;
-              delta = true; //enter into delta mode - for the remaining bits in this byte length will be 0; we will exit delta mode and enter into recovery mode on bpos==0
-            }
-            length = 0;
-          }
-        }
-
-        if (x.bpos == 0) {
-
-          // recover match after a 1-byte mismatch
-          if (isInPreRecoveryMode()) { // just exited delta mode, so we have a backup
-            //the match failed 2 bytes ago, we must increase indexBak by 2:
-            indexBak++;
-            if (lengthBak < 65535) {
-              lengthBak++;
-            }
-            if (x.buf[indexBak] == (U8)x.c4) { // match continues -> recover
-              length = lengthBak;
-              index = indexBak;
-            }
-            else { // still mismatch
-              lengthBak = indexBak = 0; // purge backup (give up)
-            }
-          }
-
-          // extend current match
-          if (length != 0) {
-            index++;
-            if (length < 65535) {
-              length++;
-            }
-            if (isInRecoveryMode() && recoveryModePos() >= MINLEN_RM) { // strong recovery -> exit tecovery mode (finalize)
-              lengthBak = indexBak = 0; // purge backup
-            }
-          }
-          delta = false;
-        }
-       /* if constexpr(false) {
-          printf("  pos %d %d  index %d  length %d  lengthBak %d  delta %d\n", blockPos, bpos, index, length, lengthBak, delta ? 1 : 0);
-        }*/
-
-      }
-
-      void registerMatch(const uint32_t pos, const uint32_t LEN) {
-        assert(pos != 0);
-        length = LEN - LEN5 + 1; // rebase
-        index = pos;
-        lengthBak = indexBak = 0;
-        expectedByte = 0;
-        delta = false;
-      }
-
-    };
-
-    bool isMatch(BlockData& x,const uint32_t pos, const uint32_t MINLEN) {
-      for (uint32_t length = 1; length <= MINLEN; length++) {
-        if (x.buf(length) != x.buf[pos - length])
-          return false;
-      }
-      return true;
-    }
-
-    void AddCandidates(BlockData& x,HashElementForMatchPositions* matches, uint32_t LEN) {
-      uint32_t i = 0;
-      while (numberOfActiveCandidates < N && i < HashElementForMatchPositions::N) {
-        uint32_t matchpos = matches->matchPositions[i];
-        if (matchpos == 0)
-          break;
-        if (isMatch(x,matchpos, LEN)) {
-          bool isSame = false;
-          //is this position already registered?
-          for (uint32_t j = 0; j < numberOfActiveCandidates; j++) {
-            MatchInfo* oldcandidate = &matchCandidates[j];
-            if (isSame = oldcandidate->index == matchpos)
-              break;
-          }
-          if (!isSame) { //don't register an already registered sequence
-            matchCandidates[numberOfActiveCandidates].registerMatch(matchpos, LEN);
-            numberOfActiveCandidates++;
-          }
-        }
-        i++;
-      }
-    }
-
-    static constexpr size_t N = 4;
-    Array<MatchInfo> matchCandidates{ N };
-    uint32_t numberOfActiveCandidates = 0;
-void update();
-public:
-    virtual ~matchModel2(){ }
-    matchModel2(BlockData& bd);
-    int p(Mixer &m,int val1,int val2);
-    int inputs();
-    int nets();
-    int netcount();
-};
-matchModel2::matchModel2(BlockData& bd) : 
-  x(bd),
-  hashtable(CMlimit(MEM()/8)),
-  stateMaps {{  28 * 512},
-             {  8 * 256 * 256 + 1},
-             {  256 * 256}},
-  cm( CMlimit(MEM()/2), nCM, 0,
-  CM_RUN1+
-  CM_RUN0+
-  CM_MAIN1+
-  CM_MAIN2+
-  CM_MAIN3+
-  CM_MAIN4+
-  CM_M12+
-  CM_M6
-  ),
-  mapL{ /* LargeStationaryMap : HashBits, Scale=64, Rate=16  */
-        { 20}, // effective bits: ~22
-  },
-  map { /* StationaryMap : BitsOfContext, InputBits, Scale=64, Rate=16  */
-        { 1 /*< leading bit */ + iCtxBits + 3 /*< length3Rm */, 1}
-  }, 
-  iCtx{15, 1, iCtxBits},
-  hashBits(ilog2(uint32_t(hashtable.size())))
-{
-  assert(isPowerOf2(hashtable.size()));
-  assert(isPowerOf2(CMlimit(MEM())));
-  iCtx.reset();
-  x.Match.bypass=false;
-  x.Match.bypassprediction=2048;
-}
-
-void matchModel2::update() {
-
-  size_t n = max(numberOfActiveCandidates, 1);
-  for (size_t i = 0; i < n; i++) {
-    MatchInfo* matchInfo = &matchCandidates[i];
-    matchInfo->update(x);
-    if (numberOfActiveCandidates != 0 && matchInfo->length == 0 && !matchInfo->delta && matchInfo->lengthBak==0) {
-      numberOfActiveCandidates--;
-      if (numberOfActiveCandidates == i)
-        break;
-      memmove(&matchCandidates[i], &matchCandidates[i + 1], (numberOfActiveCandidates - i) * sizeof(MatchInfo));
-      i--;
-    }
-  }
-
-  if( x.bpos == 0 ) {
-    uint64_t hash;
-    HashElementForMatchPositions* matches;
-
-    hash = x.cxt[LEN9];
-    matches = &hashtable[finalize64(hash, hashBits)];
-    if (numberOfActiveCandidates < N)
-      AddCandidates(x,matches, LEN9); //note: this length is *not* modelled by NormalModel
-    matches->Add(x.buf.pos);
-
-    hash = x.cxt[LEN7];
-    matches = &hashtable[finalize64(hash, hashBits)];
-    if (numberOfActiveCandidates < N)
-      AddCandidates(x,matches, LEN7); //note: this length is *not* modelled by NormalModel
-    matches->Add(x.buf.pos);
-
-    hash = x.cxt[LEN5];
-    matches = &hashtable[finalize64(hash, hashBits)];
-    if (numberOfActiveCandidates < N)
-      AddCandidates(x,matches, LEN5); //note: this length is modelled by NormalModel
-    matches->Add(x.buf.pos);
-
-    for (size_t i = 0; i < numberOfActiveCandidates; i++) {
-      matchCandidates[i].expectedByte = x.buf[matchCandidates[i].index];
-    }
-  }
-}
-
-int matchModel2::inputs()   {return 2+nCM*cm.inputs()+nST*2 +nLSM*3+nSM*2/*+3*2 */;}
-int matchModel2::nets()     {return 12;}
-int matchModel2::netcount() {return 1;}
-
-int matchModel2::p(Mixer &m,int val1=0,int val2=0) {
-  
-  update();
-
-  for( uint32_t i = 0; i < nST; i++ ) { // reset contexts
-    ctx[i] = 0;
-  }
-  
-  size_t bestCandidateIdx = 0; //default item is the first candidate, let's see if any other candidate is better
-  for (int i = 1; i < numberOfActiveCandidates; i++) {
-    if (matchCandidates[i].isBetterThan(&matchCandidates[bestCandidateIdx]))
-      bestCandidateIdx = i;
-  }
-
-  const uint32_t length = matchCandidates[bestCandidateIdx].length;
-  const uint8_t expectedByte = matchCandidates[bestCandidateIdx].expectedByte;
-  const bool isInNoMatchMode= matchCandidates[bestCandidateIdx].isInNoMatchMode();
-  const bool isInDeltaMode = matchCandidates[bestCandidateIdx].delta;
-  const bool isInPreRecoveryMode = matchCandidates[bestCandidateIdx].isInPreRecoveryMode();
-  const bool isInRecoveryMode = matchCandidates[bestCandidateIdx].isInRecoveryMode();
-  if (x.bpos == 0 && length==0 && x.Match.bypass)
-      x.Match.bypass = false; // can quit bypass mode on byte boundary only
-  const int expectedBit = length != 0 ? (expectedByte >> (7 - x.bpos)) & 1 : 0;
-  uint32_t denselength = 0; // 0..27
-  if (length != 0 && x.Match.bypass==false) {
-    if (length <= 16) {
-      denselength = length - 1; // 0..15
-    } else {
-      denselength = 12 + (min(length - 1, 63) >> 2); // 16..27
-    }
-    ctx[0] = (denselength << 9) | (expectedBit << 8) | x.c0; // 1..28*512
-    ctx[1] = ((expectedByte << 11) | (x.bpos << 8) | (U8)x.c4) + 1;
-    
-    const int sign = 2 * expectedBit - 1;
-    m.add(sign * (min(length, 32) << 5)); // +/- 32..1024
-    m.add(sign * (ilog(min(length, 65535)) << 2)); // +/-  0..1024
-  } else { // no match at all or delta mode
-    m.add(0);
-    m.add(0);
-  }
-
-  if( isInDeltaMode ) { // delta mode: helps predicting the remaining bits of a character when a mismatch occurs
-    ctx[2] = (expectedByte << 8) | x.c0;
-  }
- if ( x.Match.bypass==false){
-
-  for( uint32_t i = 0; i < nST; i++ ) {
-    const uint32_t c = ctx[i]; 
-    
-    if( c != 0 ) {
-     const int p1 = stateMaps[i].p(c,x.y);
-      const int st = stretch(p1);
-      m.add(st >> 2);
-     m.add((p1 - 2048) >> 3);
-    } else {
-      m.add(0);
-      m.add(0);
-    }
-  }
-}
-  const uint32_t lengthIlog2 = ilog2(length + 1);
-  //length=0:      lengthIlog2=0
-  //length=1..2:   lengthIlog2=1
-  //length=3..6:   lengthIlog2=2
-  //length=7..14:  lengthIlog2=3
-  //length=15..30: lengthIlog2=4
-
-  const uint8_t mode = 
-    isInNoMatchMode ? 0 :
-    isInDeltaMode ? 1 :
-    isInPreRecoveryMode ? 2 :
-    isInRecoveryMode ? 3 :
-    3 + lengthIlog2;
-  uint8_t mode3 = min(mode, 7); // 3 bits
-  
-  //bytewise contexts
-  if( x.bpos == 0 && x.Match.bypass==false) {
-    cm.set(hash(U32(length != 0 ? expectedByte : (U8)x.c4)<<3 | mode3)); //max context bits: 8+8+3 = 19
-    cm.set( hash(length != 0 ? expectedByte : ((x.c4 >> 8) & 0xff), ((U8)x.c4<<3)| mode3)); //max context bits: 8+8+8+3=27
-    
-  }
-  cm.mix(m);
-
-  //bitwise contexts
-  mapL[0].set(hash(expectedByte, x.c0, x.c4 & 0x00ffffff, mode3)); // max context bits: 8+8+24+3 = 43 bits -> hashed into ~22 bits
-  mapL[0].mix(m);
-
-  const uint32_t mCtx =
-    isInNoMatchMode ? 0 :
-    isInDeltaMode ? 1 :
-    isInPreRecoveryMode ? 2 :
-    isInRecoveryMode ? 3 :
-    4 + ((lengthIlog2 - 1) << 1 | expectedBit);
-
-  iCtx += x.y;
-  iCtx = (x.bpos << 12) | ((U8)x.c4 << 4) | min(mCtx,15); // 15 bits
-  map[0].set(iCtx() << 3 | mode3); // (max 7 bits + 1 leading bit) + 3 bits
-  map[0].mix(m);
-  
-  m.set(min(mCtx, 11),12);
-  x.Match.length3 =
-    length == 0 ? 0 :
-    isInDeltaMode ? 1 :
-    length <= 7 ? 2 : 3;
-  //x.Match.length3=(min(ilog2(length),3));
- 
-    //x.Match.byte = (length)?expectedByte:0;
-    //x.Match.bypassprediction = length==0 ? 2048 : (expectedBit==0 ? 1 : 4095);
-    //x.Match.length = length;
-  return ilog(length);
-}  
 
 
     
@@ -5918,64 +5255,7 @@ virtual ~XMLModel1(){ }
 
 
 
-static const int primes[17]={ 0, 257,251,241,239,233,229,227,223,211,199,197,193,191,181,179,173};   
-class normalModel1: public Model {
-  BlockData& x;
-  Buf& buf;
-  const int N;
-  ContextMap2   cm;
-  StateMap StateMaps[4];
-  RunContextMap /*rcm7,*/ rcm9, rcm10;
-  int inpt;
-public:
-  normalModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf), N(10), cm(CMlimit(MEM()*32), N,M_NORMAL,
-  CM_RUN1+
-  CM_RUN0+
-  CM_MAIN1+
-  CM_MAIN2+
-  CM_MAIN3+
-  CM_MAIN4+
-  CM_M12+
-  CM_M6
-  ), StateMaps{ 256, 256*256,256*256,256*256 },
-  /*rcm7(CMlimit(MEM()/(level>8?8:4)),bd),*/
-  rcm9(CMlimit(MEM()/((level>8?8:4))),bd), rcm10(CMlimit(MEM()/(level>8?4:2)),bd){
- }
- int inputs() {return 10*cm.inputs() +3+2+2+1;}
- int nets() {return 0;}
-  int netcount() {return 0;}
-  
-  
-int p(Mixer& m,int val1=0,int val2=0){  
-  if (x.bpos==0) {
-     // if (val2==-1) return 1;
-    int i;
-    if (val2==0) cm.set(x.cxt[15]);
-    for (i=1; i<=7; ++i)
-      cm.set(x.cxt[i]);
 
-    //rcm7.set(x.cxt[7]);
-    cm.set(x.cxt[9]);
-    rcm9.set(x.cxt[10]);
-    
-    rcm10.set(x.cxt[12]);
-    cm.set(x.cxt[14]);
-
-  }
-  
-  //rcm7.mix(m);
-  rcm9.mix(m);
-  rcm10.mix(m);
-  m.add((stretch(StateMaps[0].p(x.c0-1,x.y)))>>2);
-  m.add((stretch(StateMaps[1].p((x.c0-1)|(buf(1)<<8),x.y)))>>2);
-  m.add((stretch(StateMaps[2].p((x.c0-1)|(buf(1)<<8),x.y,64)))>>2);
-  m.add((stretch(StateMaps[3].p((x.c0-1)|(buf(2)<<8),x.y,64)))>>2);
-  return max(0, cm.mix(m)-(N-7)); 
-}
-  virtual ~normalModel1(){
-
- }
-};
