+++ prt/run.cpp
@@ -1,0 +1,54 @@
+#include "run.hpp"
+ 
+/////////////////////////// ContextMap /////////////////////////
+//
+// A ContextMap maps contexts to a bit histories and makes predictions
+// to a Mixer.  Methods common to all classes:
+//
+// ContextMap cm(M, C); creates using about M bytes of memory (a power
+//   of 2) for C contexts.
+// cm.set(cx);  sets the next context to cx, called up to C times
+//   cx is an arbitrary 32 bit value that identifies the context.
+//   It should be called before predicting the first bit of each byte.
+// cm.mix(m) updates Mixer m with the next prediction.  Returns 1
+//   if context cx is found, else 0.  Then it extends all the contexts with
+//   global bit y.  It should be called for every bit:
+//
+//     if (bpos==0)
+//       for (int i=0; i<C; ++i) cm.set(cxt[i]);
+//     cm.mix(m);
+//
+// The different types are as follows:
+//
+// - RunContextMap.  The bit history is a count of 0-255 consecutive
+//     zeros or ones.  Uses 4 bytes per whole byte context.  C=1.
+//     The context should be a hash.
+// - SmallStationaryContextMap.  0 <= cx < M/512.
+//     The state is a 16-bit probability that is adjusted after each
+//     prediction.  C=1.
+// - ContextMap.  For large contexts, C >= 1.  Context need not be hashed.
+
+
+
+// A RunContextMap maps a context into the next byte and a repeat
+// count up to M.  Size should be a power of 2.  Memory usage is 3M/4.
+
+  RunContextMap::RunContextMap(int m,BlockData& bd): t(m/4),x(bd) {cp=t[0]+1;}
+  void RunContextMap::set(U32 cx) {  // update count
+    if (cp[0]==0 || cp[1]!=x.buf(1)) cp[0]=1, cp[1]=x.buf(1);
+    else if (cp[0]<255) ++cp[0];
+    cp=t[cx]+1;
+  }
+  
+  int RunContextMap::p() {  // predict next bit
+    if ((cp[1]+256)>>(8-x.bpos)==x.c0)
+      return ((cp[1]>>(7-x.bpos)&1)*2-1)*ilog(cp[0]+1)*8;
+    else
+      return 0;
+  }
+  int RunContextMap::mix(Mixer& m) {  // return run length
+    m.add(p());
+    return cp[0]!=0;
+  }
+
+
